Use SourceDispatcher to route requests to async source
