+++ airbyte-cdk/python/airbyte_cdk/sources/async_cdk/abstract_source_async.py
@@ -2,15 +2,10 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
 
-import asyncio
 import logging
-import traceback
 from abc import ABC, abstractmethod
-from queue import Queue
-from threading import Lock, Thread
-from typing import Any, Callable, Dict, Iterable, Iterator, List, Mapping, MutableMapping, Optional, Tuple, TypeVar, Union
+from typing import Any, Iterable, Iterator, List, Mapping, MutableMapping, Optional, Tuple, Union
 
-import aiohttp
 from airbyte_cdk.models import (
     AirbyteConnectionStatus,
     AirbyteMessage,
@@ -28,71 +23,8 @@ from airbyte_cdk.sources.async_cdk.streams.core_async import AsyncStream
 from airbyte_cdk.sources.streams.core import StreamData
 from airbyte_cdk.sources.streams.http.http import HttpStream
 from airbyte_cdk.sources.utils.record_helper import stream_data_to_airbyte_message
-from airbyte_cdk.sources.utils.schema_helpers import InternalConfig, split_config
-from airbyte_cdk.utils.event_timing import create_timer
+from airbyte_cdk.sources.utils.schema_helpers import InternalConfig
 from airbyte_cdk.utils.stream_status_utils import as_airbyte_message as stream_status_as_airbyte_message
-from airbyte_cdk.utils.traced_exception import AirbyteTracedException
-
-T = TypeVar("T")
-DEFAULT_SESSION_LIMIT = 10000
-DEFAULT_TIMEOUT = None
-
-
-class Sentinel:
-    def __init__(self, name: str):
-        self.name = name
-
-
-class SourceReader(Iterator):
-    def __init__(self, logger: logging.Logger, queue: Queue, sentinels: Dict[str, Sentinel], reader_fn: Callable, *args: Any):
-        self.logger = logger
-        self.queue = queue
-        self.sentinels = sentinels
-        self.reader_fn = reader_fn
-        self.reader_args = args
-        self.sessions: Dict[str, aiohttp.ClientSession] = {}
-
-        self.thread = Thread(target=self._start_reader_thread)
-        self.thread.start()
-
-    def _start_reader_thread(self):
-        asyncio.run(self.reader_fn(*self.reader_args))
-
-    def __next__(self):
-        loop = asyncio.get_event_loop()
-        try:
-            item = self.queue.get(timeout=DEFAULT_TIMEOUT)
-            if isinstance(item, Exception):
-                self.logger.error(f"An error occurred in the async thread: {item}")
-                self.thread.join()
-                raise item
-            if isinstance(item, Sentinel):
-                # Sessions can only be closed once items in the stream have all been dequeued
-                if session := self.sessions.pop(item.name, None):
-                    loop.create_task(session.close())  # TODO: this can be done better
-                try:
-                    self.sentinels.pop(item.name)
-                except KeyError:
-                    raise RuntimeError(f"The sentinel for stream {item.name} was already dequeued. This is unexpected and indicates a possible problem with the connector. Please contact Support.")
-                if not self.sentinels:
-                    self.thread.join()
-                    raise StopIteration
-                else:
-                    return self.__next__()
-            else:
-                return item
-        finally:
-            loop.create_task(self.cleanup())
-
-    def drain(self):
-        while not self.queue.empty():
-            yield self.queue.get()
-        self.thread.join()
-
-    async def cleanup(self):
-        # pass
-        for session in self.sessions.values():
-            await session.close()
 
 
 class AsyncAbstractSource(AbstractSource, ABC):
@@ -101,13 +33,8 @@ class AsyncAbstractSource(AbstractSource, ABC):
     in this class to create an Airbyte Specification compliant Source.
     """
 
-    def __init__(self):
-        super().__init__()
-        self.queue = Queue(10000)
-        self.session_limit = DEFAULT_SESSION_LIMIT
-
     @abstractmethod
-    def check_connection(self, logger: logging.Logger, config: Mapping[str, Any]) -> Tuple[bool, Optional[Any]]:
+    async def check_connection(self, logger: logging.Logger, config: Mapping[str, Any]) -> Tuple[bool, Optional[Any]]:
         """
         :param logger: source logger
         :param config: The user-provided configuration as specified by the source's spec.
@@ -129,132 +56,28 @@ class AsyncAbstractSource(AbstractSource, ABC):
         return AirbyteConnectionStatus(status=Status.SUCCEEDED)
 
     @abstractmethod
-    def streams(self, config: Mapping[str, Any]) -> List[AsyncStream]:
+    async def streams(self, config: Mapping[str, Any]) -> List[AsyncStream]:
         """
         :param config: The user-provided configuration as specified by the source's spec.
         Any stream construction related operation should happen here.
         :return: A list of the streams in this source connector.
         """
 
-    def read(
+    async def read(
         self,
         logger: logging.Logger,
         config: Mapping[str, Any],
         catalog: ConfiguredAirbyteCatalog,
         state: Optional[Union[List[AirbyteStateMessage], MutableMapping[str, Any]]] = None,
     ) -> Iterator[AirbyteMessage]:
-        """Implements the Read operation from the Airbyte Specification. See https://docs.airbyte.com/understanding-airbyte/airbyte-protocol/."""
-        logger.info(f"Starting syncing {self.name}")
-        config, internal_config = split_config(config)
-        stream_instances = {s.name: s for s in self.streams(config)}
-        state_manager = ConnectorStateManager(stream_instance_map=stream_instances, state=state)
-        self._stream_to_instance_map = stream_instances
-        self._assert_streams(catalog, stream_instances)
+        """
+        Implements the Read operation from the Airbyte Specification. See https://docs.airbyte.com/understanding-airbyte/airbyte-protocol/.
 
-        n_records = 0
-        with create_timer(self.name) as timer:
-            for record in self._do_read(catalog, stream_instances, timer, logger, state_manager, internal_config):
-                n_records += 1
-                yield record
+        This method is not used when the AsyncSource is used in conjunction with the AsyncSourceDispatcher.
+        """
+        ...
 
-        print(f"_______________________-ASYNCIO SOURCE N RECORDS == {n_records}")
-        logger.info(f"Finished syncing {self.name}")
-
-    def _do_read(self, catalog: ConfiguredAirbyteCatalog, stream_instances: Dict[str, AsyncStream], timer: Any, logger: logging.Logger, state_manager: ConnectorStateManager, internal_config: InternalConfig):
-        streams_in_progress_sentinels = {s.stream.name: Sentinel(s.stream.name) for s in catalog.streams if s.stream.name in stream_instances}
-        if not streams_in_progress_sentinels:
-            return
-        self.reader = SourceReader(logger, self.queue, streams_in_progress_sentinels, self._read_streams, catalog, stream_instances, timer, logger, state_manager, internal_config)
-        for record in self.reader:
-            yield record
-        print("DRAINING __________________________")
-        for record in self.reader.drain():
-            if isinstance(record, Exception):
-                raise record
-            yield record
-
-    def _assert_streams(self, catalog: ConfiguredAirbyteCatalog, stream_instances: Dict[str, AsyncStream]):
-        for configured_stream in catalog.streams:
-            stream_instance = stream_instances.get(configured_stream.stream.name)
-            if not stream_instance:
-                if not self.raise_exception_on_missing_stream:
-                    return
-                raise KeyError(
-                    f"The stream {configured_stream.stream.name} no longer exists in the configuration. "
-                    f"Refresh the schema in replication settings and remove this stream from future sync attempts."
-                )
-
-    async def _read_streams(self, catalog: ConfiguredAirbyteCatalog, stream_instances: Dict[str, AsyncStream], timer: Any, logger: logging.Logger, state_manager: ConnectorStateManager, internal_config: InternalConfig):
-        pending_tasks = set()
-        n_started, n_streams = 0, len(catalog.streams)
-        streams_iterator = iter(catalog.streams)
-        exceptions = False
-
-        while (pending_tasks or n_started < n_streams) and not exceptions:
-            while len(pending_tasks) < self.session_limit and (configured_stream := next(streams_iterator, None)):
-                if configured_stream is None:
-                    break
-                # stream_instance = stream_instances.get("Account")  # TODO
-                stream_instance = stream_instances.get(configured_stream.stream.name)
-                stream = stream_instances.get(configured_stream.stream.name)
-                self.reader.sessions[configured_stream.stream.name] = await stream.ensure_session()
-                pending_tasks.add(asyncio.create_task(self._do_async_read_stream(configured_stream, stream_instance, timer, logger, state_manager, internal_config)))
-                n_started += 1
-
-            done, pending_tasks = await asyncio.wait(pending_tasks, return_when=asyncio.FIRST_COMPLETED)
-
-            for task in done:
-                if exc := task.exception():
-                    for remaining_task in pending_tasks:
-                        await remaining_task.cancel()
-                    self.queue.put(exc)
-                    exceptions = True
-
-    async def _do_async_read_stream(self, configured_stream: ConfiguredAirbyteStream, stream_instance: AsyncStream, timer: Any, logger: logging.Logger, state_manager: ConnectorStateManager, internal_config: InternalConfig):
-        try:
-            await self._async_read_stream(configured_stream, stream_instance, timer, logger, state_manager, internal_config)
-        finally:
-            self.queue.put(Sentinel(configured_stream.stream.name))
-
-    async def _async_read_stream(self, configured_stream: ConfiguredAirbyteStream, stream_instance: AsyncStream, timer: Any, logger: logging.Logger, state_manager: ConnectorStateManager, internal_config: InternalConfig):
-        try:
-            timer.start_event(f"Syncing stream {configured_stream.stream.name}")
-            stream_is_available, reason = await stream_instance.check_availability(logger, self)
-            if not stream_is_available:
-                logger.warning(f"Skipped syncing stream '{stream_instance.name}' because it was unavailable. {reason}")
-                return
-            logger.info(f"Marking stream {configured_stream.stream.name} as STARTED")
-            self.queue.put(stream_status_as_airbyte_message(configured_stream.stream, AirbyteStreamStatus.STARTED))
-            async for record in self._read_stream(
-                logger=logger,
-                stream_instance=stream_instance,
-                configured_stream=configured_stream,
-                state_manager=state_manager,
-                internal_config=internal_config,
-            ):
-                self.queue.put(record)
-            logger.info(f"Marking stream {configured_stream.stream.name} as STOPPED")
-            self.queue.put(stream_status_as_airbyte_message(configured_stream.stream, AirbyteStreamStatus.COMPLETE))
-        except AirbyteTracedException as e:
-            self.queue.put(stream_status_as_airbyte_message(configured_stream.stream, AirbyteStreamStatus.INCOMPLETE))
-            raise e
-        except Exception as e:
-            for message in self._emit_queued_messages():
-                self.queue.put(message)
-            logger.exception(f"Encountered an exception while reading stream {configured_stream.stream.name}")
-            logger.info(f"Marking stream {configured_stream.stream.name} as STOPPED")
-            self.queue.put(stream_status_as_airbyte_message(configured_stream.stream, AirbyteStreamStatus.INCOMPLETE))
-            display_message = await stream_instance.get_error_display_message(e)
-            if display_message:
-                raise AirbyteTracedException.from_exception(e, message=display_message)
-            else:
-                raise e
-        finally:
-            timer.finish_event()
-            logger.info(f"Finished syncing {configured_stream.stream.name}")
-            # logger.info(timer.report())  # TODO - this is causing scenario-based test failures
-
-    async def _read_stream(
+    async def read_stream(
         self,
         logger: logging.Logger,
         stream_instance: AsyncStream,
