+++ airbyte-cdk/python/unit_tests/sources/streams/concurrent/scenarios/thread_based_concurrent_stream_source_builder.py
@@ -18,8 +18,8 @@ from airbyte_cdk.sources.streams.concurrent.partitions.partition import Partitio
 from airbyte_cdk.sources.streams.concurrent.partitions.partition_generator import PartitionGenerator
 from airbyte_cdk.sources.streams.concurrent.partitions.record import Record
 from airbyte_cdk.sources.streams.core import StreamData
-from airbyte_cdk.sources.utils.slice_logger import SliceLogger
 from airbyte_protocol.models import ConfiguredAirbyteStream
+from unit_tests.sources.scenario_based.helpers import NeverLogSliceLogger
 from unit_tests.sources.scenario_based.scenario_builder import SourceBuilder
 
 
@@ -134,8 +134,3 @@ class ConcurrentSourceBuilder(SourceBuilder[ConcurrentCdkSource]):
 class AlwaysAvailableAvailabilityStrategy(AbstractAvailabilityStrategy):
     def check_availability(self, logger: logging.Logger) -> StreamAvailability:
         return StreamAvailable()
-
-
-class NeverLogSliceLogger(SliceLogger):
-    def should_log_slice_message(self, logger: logging.Logger) -> bool:
-        return False
