+++ airbyte-integrations/connectors/source-salesforce/main.py
@@ -10,16 +10,18 @@ from typing import List
 
 from airbyte_cdk.entrypoint import AirbyteEntrypoint, launch
 from airbyte_cdk.models import AirbyteErrorTraceMessage, AirbyteMessage, AirbyteTraceMessage, TraceType, Type
-from source_salesforce import SourceSalesforce
+from source_salesforce import SalesforceSourceDispatcher, SourceSalesforce
 
 
 def _get_source(args: List[str]):
     catalog_path = AirbyteEntrypoint.extract_catalog(args)
     config_path = AirbyteEntrypoint.extract_config(args)
     try:
-        return SourceSalesforce(
-            SourceSalesforce.read_catalog(catalog_path) if catalog_path else None,
-            SourceSalesforce.read_config(config_path) if config_path else None,
+        return SalesforceSourceDispatcher(
+            SourceSalesforce(
+                SourceSalesforce.read_catalog(catalog_path) if catalog_path else None,
+                SourceSalesforce.read_config(config_path) if config_path else None,
+            )
         )
     except Exception as error:
         print(
@@ -39,7 +41,8 @@ def _get_source(args: List[str]):
 
 
 if __name__ == "__main__":
-    _args = ['read', '--config', 'secrets/config.json', '--catalog', 'integration_tests/configured_catalog.json']
+    # _args = ['read', '--config', 'secrets/config.json', '--catalog', 'integration_tests/configured_catalog.json']
+    _args = sys.argv[1:]
     source = _get_source(_args)
     if source:
         launch(source, _args)
